+++ src/ui/lv_conf.h
@@ -22,7 +22,9 @@
 #define LV_USE_USER_DATA 1
 
 #define LV_FONT_MONTSERRAT_14 1
-#define LV_FONT_DEFAULT &lv_font_montserrat_14
+#define LV_FONT_MONTSERRAT_16 1
+#define LV_FONT_MONTSERRAT_20 1
+#define LV_FONT_DEFAULT &lv_font_montserrat_16
 #define LV_USE_THEME_DEFAULT 1
 
 #endif  // LV_CONF_H